+++ src/test/kotlin/nl/dirkgroot/structurizr/dsl/parser/SingleLineElementsTest.kt
@@ -1,12 +1,12 @@
 package nl.dirkgroot.structurizr.dsl.parser
 
 import nl.dirkgroot.structurizr.dsl.lexer.KEYWORDS_WITH_BLOCKS
-import nl.dirkgroot.structurizr.dsl.support.StructurizrDSLParserTest
+import nl.dirkgroot.structurizr.dsl.support.StructurizrDSLCodeInsightTest
 import org.junit.jupiter.api.DynamicTest.dynamicTest
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestFactory
 
-class SingleLineElementsTest : StructurizrDSLParserTest() {
+class SingleLineElementsTest : StructurizrDSLCodeInsightTest() {
     @TestFactory
     fun `without argument`() = KEYWORDS_WITH_BLOCKS.map { (keyword, _) ->
         dynamicTest(keyword) {